requirements.txt
dronekit
pymavlink
python-dotenv